EyeOfRa/scripts/pipeline/Ipy/requirements.txt
dask==2022.4.0
dask_ml==2022.1.22
h5py==3.6.0
joblib==1.1.0
numpy==1.21.5
opencv_python==4.5.5.64
Pillow==9.1.0
scikit_learn==1.0.2
scikit_optimize==0.9.0
scipy==1.8.0
snakemake==7.3.8
nbformat==5.0.8
